The comparison between Simulation and Measured has been implemented
The comparison between simulated and masured data now is performed by
a naif implementation of a single plane module-loader.
If you want to make the comparison, now you have to lunch the program
with the pasrser

-ref Freq1,Plane1EFreq2,Plane2E...

The E letter split the different planes while the comma split
the frequency of the plane. I know, it's a very naif implementation.
A better solution would be represented by the sequent parser.

-ref Plane1,Freq1,Freq2,Freq3EPlane2,Freq1,Freq2,Freq3

Maybe it is the way...